explicit 64-bit instead of size_t

+++ src/Serialize.h
@@ -10,7 +10,7 @@ void serialize(std::ostream& stream, bool value);
 template<typename T>
 void serialize(std::ostream& stream, const std::vector<T>& value)
 {
-	serialize(stream, value.size());
+	serialize(stream, (uint64_t)value.size());
 	for (size_t i = 0; i < value.size(); i++)
 	{
 		serialize(stream, value[i]);
@@ -23,7 +23,7 @@ void deserialize(std::istream& stream, bool& value);
 template<typename T>
 void deserialize(std::istream& stream, std::vector<T>& value)
 {
-	size_t size = 0;
+	uint64_t size = 0;
 	deserialize(stream, size);
 	value.resize(size);
 	for (size_t i = 0; i < value.size(); i++)
